refactor: split robust library
